projeto-transformacao-digital/requirements.txt
functions-framework==3.*
google-cloud-bigquery==3.*
google-cloud-storage==2.*
pytest~=7.4.3
cloudevents~=1.10.0